pyproject.toml
[project]
name = "nonebot-plugin-nonememe"
version = "0.2.0.post3"
description = "The daily life of the NoneBot group members"
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-plugin-send-anything-anywhere>=0.2.7",
    "nonebot-plugin-apscheduler>=0.3.0",
    "pydantic>=1.10.4,<2",
    "httpx>=0.24.1",
    "json5>=0.9.14",
    "anyio>=3.7.1",
    "typing-extensions>=4.7.1",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-nonememe"

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
